Bot is full ready

+++ src/scenes/admin.scene.ts
@@ -15,7 +15,6 @@ Adminscene.step(async(ctx)=>{
 
     const id = (ctx as any).from.id;
 
-    (ctx as any).session.chatId = ctx.chat?.id;
 
     if(id!=APPLICATION.admin_id){
         await ctx.reply(AdminSceneMessages.forbidden)
@@ -61,18 +60,24 @@ Adminscene.wait("give-add-ingore-add").on("callback_query:data",async(ctx)=>{
     const query = ctx.callbackQuery.data.toLowerCase().trim();
     const id=query.split("_")[1]
     const deleteid=query.split("_")[2]
-
+    const post=await postservice.getPosts(NeededCount.ONE,id)
     if(query.startsWith(SomeNeccessaryMessages.accept)){
+      if(post){
         await postservice.updatePost(id)
         await ctx.api.deleteMessage((ctx as any).session.chatId ,+deleteid)
-        const post=await postservice.getPosts(NeededCount.ONE,id)
         const format=await formatservice.createTemplate(post)
         await ctx.api.sendMessage(APPLICATION.admin_id,SomeNeccessaryMessages.good)
         await ctx.api.sendMessage(APPLICATION.channel,format)
+      }else{
+        await ctx.api.sendMessage(APPLICATION.admin_id,SomeNeccessaryMessages.ignore)
+      }
     }else if(query.startsWith(SomeNeccessaryMessages.reject)){
-      await ctx.api.deleteMessage((ctx as any).session.chatId ,+deleteid)
+      if(post){
         await postservice.deletePost(id)
         await ctx.api.sendMessage(APPLICATION.admin_id,SomeNeccessaryMessages.bad)
+      }else{
+        await ctx.api.sendMessage(APPLICATION.admin_id,SomeNeccessaryMessages.ignore)
+      }
     }else if(query==SomeNeccessaryMessages.back){
         await ctx.reply(AdminSceneMessages.getMainMenu)
         return ctx.scene.exit()
